Only mount shared DB drive if already mounting
This shifts the default behavior for VVV to non-persisting
databases. If an existing VM is already storing database data on
the guest machine, drives will continue to mount. If you remove
this data or start fresh with VVV, the MySQL directory will not
be stored on your computer.

This addresses the pain that is #287 and moves to a model that
(to me) makes more sense at all.

We'll probably want to think of ways we can avoid data loss by
making it easy to backup data on a regular basis as required.