Install the mainline version of Nginx
In the world of Nginx, mainline can be considered the most feature
rich and the most stable. The "stable" release receives only bug
updates. A good explanation of this was posted by the nginx project
after the initial 1.7.x release from 1.6.x

http://nginx.com/blog/nginx-1-6-1-7-released/

I've used nginx 1.7.x in production since that day and have not had
any troubles. I think we should be good to move to mainline as our
default provisioning.