Update Composer packages on every provision
In #129, we moved from a composer.json to a global configuration
for packages such as phpunit. In the current state, changing the
version numbers for these packages only modifies new boxes.

Instead, we can check for and install Composer separately from
a self update and a global update. This allows version changes
to be applied with a standard `vagrant provision`.